feat(@lexical/devtools-core): Added core devtools package
