+++ packages/lexical-website/docs/packages/lexical-devtools-core.md
@@ -1,0 +1,8 @@
+---
+title: ''
+sidebar_position: 20
+sidebar_label: '@lexical/devtools-core'
+cache_reset: 1
+---
+
+{@import ../../../lexical-devtools-core/README.md}
