+++ packages/lexical/src/index.ts
@@ -15,6 +15,7 @@ export type {
   CreateEditorArgs,
   EditableListener,
   EditorConfig,
+  EditorSetOptions,
   EditorThemeClasses,
   HTMLConfig,
   Klass,
